pyproject.toml
[tool.poetry]
name = "drf-lookup"
version = "0.1.0"
description = "Helper for retrieve select options of drf fields and django-filter filters."
authors = ["Grigory Mishchenko <grishkokot@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "drf_lookup"}]
repository = "https://github.com/kindlycat/drf-lookup"
keywords = ["drf", "django-filter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.15.1"
django-filter = "^24.2"

[tool.poetry.group.dev.dependencies]
ipython = "^8.24"
ipdb = "^0.13.13"
pytest = "^8.1.2"
ruff = "^0.4.2"
pytest-django = "^4.8.0"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
sphinx = "^7.3.7"
furo = "^2024.4.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "docs",
]

line-length = 79
indent-width = 4

target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100", "D101", "ANN101", "D102", "EXE001", "D107", "ANN003", "ANN001",
    "ARG002", "ANN002", "D103", "D106", "D104", "ANN201", "COM812", "ISC001",
    "ANN401", "ANN102", "TRY003", "EM102", "EM101", "RET503", "PLR0913"
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101", "S106", "RUF012", "D105", "PLR2004", "FBT003", "ANN202", "ANN206",
    "PT004", "DJ001",
]

[tool.ruff.lint.isort]
section-order = [
    "future", "standard-library", "third-party", "first-party", "local-folder"
]
lines-after-imports = 2
combine-as-imports = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -v"

[tool.coverage.run]
source = ["drf_lookup"]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
]
